// frontend/src/types.ts
export enum ViewState {
  DASHBOARD = 'dashboard',
  DATASETS = 'datasets',
  DATA_AUGMENTATION = 'data_augmentation',
  MODEL_BUILDER = 'model_builder',
  TRAINING = 'training',
  INFERENCE = 'inference',
  SETTINGS = 'settings'
}